Fix integer overflow in SortedRanges.invertOnNew.  Fixes #666.

+++ DB/src/main/java/io/deephaven/db/v2/utils/sortedranges/SortedRanges.java
@@ -2260,7 +2260,7 @@ public abstract class SortedRanges extends RefCountedCow<SortedRanges> implement
         long end = rit.currentRangeEnd();
         long packedStart = pack(start);
         int i = 0;
-        int pos = 0;
+        long pos = 0;
         long data = packedGet(i);
         boolean neg = false;
         long pendingStart = -1;
